+++ tests/metadata/imported_metadata_test.py
@@ -98,3 +98,70 @@ def test_importing_metadata_with_data_path_and_data_string():
     assert str(error.value) == (
         "Both data path and data string given."
     )
+
+
+@pytest.mark.parametrize(
+    "name,metadata_type,metadata_format,other_format,format_version",
+    [
+        (
+            "valid_ead3.xml",
+            metadata.MetadataType.DESCRIPTIVE,
+            metadata.MetadataFormat.OTHER,
+            "EAD3",
+            "1.1.1"
+        ),
+        (
+            "invalid_dc.xml",
+            metadata.MetadataType.DESCRIPTIVE,
+            metadata.MetadataFormat.DC,
+            None,
+            "2008"
+        )
+    ]
+)
+def test_import_metadata_from_path(
+        name, metadata_type, metadata_format, other_format, format_version):
+    """
+    Test that automatically importing metadata from an XML document produces
+    the expected ImportedMetadata
+    """
+    md = metadata.ImportedMetadata.from_path(
+        Path("tests/data") / name
+    )
+
+    assert md.metadata_type == metadata_type
+    assert md.metadata_format == metadata_format
+    assert md.other_format == other_format
+    assert md.format_version == format_version
+
+    assert md.data_path == (Path("tests/data") / name).resolve()
+
+
+def test_import_metadata_from_string():
+    """
+    Test that automatically importing metadata from an XML string produces
+    the expected ImportedMetadata
+    """
+    md = metadata.ImportedMetadata.from_string(
+        Path("tests/data/valid_ead3.xml").read_bytes()
+    )
+
+    assert md.metadata_type == metadata.MetadataType.DESCRIPTIVE
+    assert md.metadata_format == metadata.MetadataFormat.OTHER
+    assert md.other_format == "EAD3"
+    assert md.format_version == "1.1.1"
+
+    assert b"<ead " in md.data_string
+
+
+def test_import_metadata_from_string_unknown_schema():
+    """
+    Test that automatically importing metadata from an unrecognized XML
+    document results in an exception
+    """
+    with pytest.raises(ValueError) as exc:
+        metadata.ImportedMetadata.from_path(
+            Path("tests/data/imported_metadata.xml")
+        )
+
+    assert "Could not recognize the metadata schema" in str(exc.value)