+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Added
 ^^^^^
 - Add linked metadata to structural maps and digital objects automatically
 - Add ```bundle_metadata``` method to ```StructuralMapDiv``` which will move shared metadata entries to the structural map div from child digital objects and structural map divs recursively.
+- Add ``ImportedMetadata.from_path`` and ``ImportedMetadata.from_string`` to automatically detect external XML metadata and create a corresponding ``ImportedMetadata`` instance.
 
 Changed
 ^^^^^^^